app/requirements/base.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
ujson==5.5.0
python-dotenv==0.21.0
sqlmodel==0.0.8
asyncpg==0.27.0
alembic==1.8.1
sqlalchemy-utils==0.38.3
passlib==1.7.4